+++ packages/yourloops/lib/metrics.ts
@@ -31,6 +31,13 @@ import bows from "bows";
 
 import config from "./config";
 
+interface ITimerMetric {
+  name: string;
+  startTime: number;
+  duration: number;
+  result: string;
+}
+
 const log = bows("Metrics");
 let metricsEnabled = false;
 
@@ -87,6 +94,11 @@ function sendMetrics(eventName: string, properties?: unknown): void {
       matomoPaq.push(["resetUserId"]);
     } else if (eventName === "setDocumentTitle" && typeof properties === "string") {
       matomoPaq.push(["setDocumentTitle", properties]);
+    } else if (eventName === "trackPageView") {
+      matomoPaq.push(["trackPageView"]);
+    } else if (eventName === "timer") {
+      const props = properties as ITimerMetric;
+      matomoPaq.push(["trackEvent", eventName, props.name, JSON.stringify(props)]);
     } else if (typeof properties === "undefined") {
       matomoPaq.push(["trackEvent", eventName, "n/a"]);
     } else {
@@ -100,5 +112,20 @@ function sendMetrics(eventName: string, properties?: unknown): void {
     logUnknownMetricsConfiguration();
   }
 }
+
+// Quick & dirty metrics timer
+
+sendMetrics.timers = new Map<string, number>();
+sendMetrics.startTimer = (name: string) => {
+  sendMetrics.timers.set(name, Date.now());
+};
+sendMetrics.endTimer = (name: string, result: string) => {
+  const startTime = sendMetrics.timers.get(name);
+  if (_.isNumber(startTime)) {
+    sendMetrics.timers.delete(name);
+    const duration = Date.now() - startTime;
+    sendMetrics("timer", { name, startTime, duration, result });
+  }
+};
 
 export default sendMetrics;