+++ templates/matomo.js
@@ -2,7 +2,8 @@ var _paq = window._paq || [];
 /* tracker methods like "setCustomDimension" should be called before "trackPageView" */
 _paq.push(['setDoNotTrack', true]);
 _paq.push(['requireConsent']);
-_paq.push(['trackPageView']);
+_paq.push(['enableHeartBeatTimer']);
+_paq.push(['setSecureCookie', true]);
 (function () {
     const u = '//localhost:8091/';
     const id = 1;